moved setInterval into gallery
