+++ src/components/Gallery/setInterval.ts
